Reorganized listeners to follow the upstream structure
