+++ DependencyInjection/Configuration.php
@@ -101,11 +101,6 @@ final class Configuration implements ConfigurationInterface
                     ->info('Whether to enable the implicit grant')
                     ->defaultTrue()
                 ->end()
-                ->scalarNode('auth_code_ttl')
-                    ->info("How long the issued auth code should be valid for.\nThe value should be a valid interval: http://php.net/manual/en/dateinterval.construct.php#refsect1-dateinterval.construct-parameters")
-                    ->cannotBeEmpty()
-                    ->defaultValue('PT10M')
-                ->end()
                 ->scalarNode('authorization_strategy')
                     ->isRequired()
                     ->info("What strategy should be used to authorize user.\nService must implement AuthorizationDecisionStrategy interface")
